requirements.txt
pandas
streamlit
Pillow
fake-useragent